Fix overflow bug for large archives.
Fix: Changed the offset stored in SSAChunkMeta structs, including the FFI variants, to store the offset as u64 vice u32 to prevent overflow.

Fix: Adjust cast in create_file_manifest_and_chunks function to u64 due to offset now being u64 vice u32.

+++ crates/lib_sprite_shrink/src/ffi/ffi_structs.rs
@@ -249,7 +249,7 @@ pub struct FFISerializedOutputU128 {
 #[derive(Clone)]
 pub struct FFISSAChunkMetaU64 {
     pub hash: u64,
-    pub offset: u32,
+    pub offset: u64,
     pub length: u32,
 }
 
@@ -258,7 +258,7 @@ pub struct FFISSAChunkMetaU64 {
 #[derive(Clone)]
 pub struct FFISSAChunkMetaU128 {
     pub hash: u128,
-    pub offset: u32,
+    pub offset: u64,
     pub length: u32,
 }
